chapter08/requirements.txt
# Chapter 8 — Data Analysis and Reasoning Agents
# Book: "Agents" by Imran Ahmad (Packt, 2026)
#
# Install: pip install -r requirements.txt
# Note: torch and transformers are optional (for Section 8.2.5 NLI demo).
#        The notebook runs fully without them in Simulation Mode.

# ── Core (required) ──────────────────────────────────
python-dotenv>=1.1.0
numpy>=1.26.0,<3.0.0
pandas>=2.2.0,<3.0.0
matplotlib>=3.9.0,<4.0.0
statsmodels>=0.14.0,<1.0.0
openai>=1.30.0,<2.0.0

# ── Optional: NLI Demo (Section 8.2.5) ──────────────
# Uncomment these if you want to run the BART-MNLI model locally.
# Requires ~1.6 GB download on first run.
# transformers>=4.44.0,<5.0.0
# torch>=2.4.0,<3.0.0

# ── Jupyter ──────────────────────────────────────────
notebook>=7.2.0
ipykernel>=6.29.0
